Adds tree benchmark (#11)
The main idea of the benchmark is to test how the state (both read and update) is handled in the presence of
multiple resumptions. During the exploration of tree, the state is constantly updated. The formula for the update
is somewhat arbitrary, the main idea is to have a non-commutative, non-associative function
that does not commute with `max` so that native implementation can't do anything too smart.
Modular arithmetic is used to prevent any overflow issues, while `abs` is used to make it as portable as
possible (The behaviour of `mod` operator in Haskell and OCaml is different when the first argument is negative).